requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==4.2.20
distro==1.9.0
fonttools==4.57.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
kiwisolver==1.4.8
matplotlib~=3.9.4
numpy==2.2.5
openai~=1.76.0
packaging==25.0
pillow==11.2.1
pip==25.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
 
plaid-python~=30.0.0
dotenv~=0.9.9